Add settings menu #24
