+++ src/main/resources/com/change_vision/astah/extension/plugin/script/messages.properties
@@ -29,7 +29,7 @@ action.clear_console.tooltip=Clear console
 action.open_sample_script.label=Sample Scripts ...
 action.open_api_guide.label=Astah API User Guide ...
 action.open_api_reference.label=Astah API Reference ...
-action.config.label=Config
+action.config.label=Settings
 
 #uri
 sample_script_uri=https://astahblog.com/sample-scripts/
